# config/openApiDocumentation.yaml
openapi: 3.0.0
info:
  description: |
    API desarrollada por el equipo de 
    [PaperCube](http://papercube.pe) para la implementación
    del ERP de la empresa Max Import.
  version: "1.0.0"
  title: Max Import ERP
  # termsOfService: 'http://swagger.io/terms/'
  contact:
    email: soluciones@papercube.pe
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: http://localhost:3000
    description: Local Server
  - url: http://maximport-backend.q3d2pmiqsz.us-east-1.elasticbeanstalk.com
    description: Development Server
  - url: POR DEFINIR
    description: Production Server
  - url: https://virtserver.swaggerhub.com/PaperCube/max-import-papercube/1.0.0
    description: SwaggerHub API Auto Mocking
...
paths:
...
  "/supplies":
    post:
      tags:
        - supplies
      summary: Crear un nuevo abastecimiento
      operationId: addSupply
      responses:
        "201":
          description: Supply created
        "400":
          description: Bad request
      requestBody:
        $ref: "#/components/requestBodies/Supply"
    get:
      tags:
        - supplies
      summary: Listar todos los abastecimientos. Ordenado por fecha de creación. Existen filtros opcionales.
      operationId: listSupplies
      parameters:
        - $ref: "#/components/parameters/fromDate"
        - $ref: "#/components/parameters/toDate"
        - $ref: "#/components/parameters/pagePagination"
        - $ref: "#/components/parameters/pageSizePagination"
      responses:
        "200":
          description: Supplies found
...
components:
...
  parameters:
...
    fromDate:
      name: from
      in: query
      description: Fecha inicial de la consulta "YYYY-MM-DD". El valor por defecto es 1 semana antes del dia actual
      required: false
      schema:
        type: string
        format: date
    toDate:
      name: to
      in: query
      description: Fecha final de la consulta "YYYY-MM-DD". El valor por defecto es el dia actual
      required: false
      schema:
        type: string
        format: date
    pagePagination:
      name: page
      in: query
      description: Numero de pagina a consultar. Indexado en 1
      required: false
      schema:
        type: integer
        format: int32
    pageSizePagination:
      name: pageSize
      in: query
      description: Cantidad de items por pagina
      required: false
      schema:
        type: integer
        format: int32
  requestBodies:
...
    Supply:
      content:
        application/json:
          schema:
            type: object
            properties:
              providerId:
                type: integer
                format: int32
                description: Id del proveedor
              warehouseId:
                type: integer
                format: int32
                description: Id del almacen destino
              observations:
                type: string
              suppliedProducts:
                type: array
                items:
                  type: object
                  properties:
                    productId:
                      type: integer
                      format: int32
                      description: Id del producto
                    boxSize:
                      type: integer
                      format: int32
                      description: Cantidad de unidades en la caja
                    quantity:
                      type: integer
                      format: int32
                      description: Cantidad de cajas a recibir
                description: Se puede tener elementos del mismo productId pero distinto boxSize
            example:
              providerId: 1
              warehouseId: 1
              observations: "Ninguna"
              suppliedProducts:
                [
                  { productId: 1, boxSize: 20, quantity: 100 },
                  { productId: 1, boxSize: 50, quantity: 20 },
                  { productId: 2, boxSize: 50, quantity: 10 },
                ]
            required:
              - providerId
              - warehouseId
              - suppliedProducts
      required: true
...